+++ src/app/api/booking/[id]/route.ts
@@ -1,7 +1,8 @@
-import type { NextApiRequest, NextApiResponse } from "next";
+import type { NextApiResponse } from "next";
 import bookingService from "../../../../services/booking";
 import { IBooking } from "../../../../models/Booking";
 import dbConnect from "../../../../lib/dbConnect";
+import { NextRequest } from "next/server";
 
 type ResponseData = {
   success: boolean;
@@ -9,10 +10,11 @@ type ResponseData = {
   message?: string;
 };
 
-export async function GET(req: NextApiRequest, res: NextApiResponse<ResponseData>) {
+export async function GET(req: NextRequest, res: NextApiResponse<ResponseData>) {
   await dbConnect();
 
-  const { id } = req.query;
+  const data = req.nextUrl.searchParams;
+  const id = data.get("id");
 
   if (typeof id !== "string") {
     return res.status(400).json({ success: false, message: "Invalid ID" });
@@ -29,17 +31,18 @@ export async function GET(req: NextApiRequest, res: NextApiResponse<ResponseData
   }
 }
 
-export async function PUT(req: NextApiRequest, res: NextApiResponse<ResponseData>) {
+export async function PUT(req: NextRequest, res: NextApiResponse<ResponseData>) {
   await dbConnect();
 
-  const { id } = req.query;
+  const data = req.nextUrl.searchParams;
+  const id = data.get("id");
 
   if (typeof id !== "string") {
     return res.status(400).json({ success: false, message: "Invalid ID" });
   }
 
   try {
-    const booking = await bookingService.updateBooking(id, req.body);
+    const booking = await bookingService.updateBooking(id, await req.json());
     if (!booking) {
       return res.status(404).json({ success: false, message: "Booking not found" });
     }
@@ -49,11 +52,12 @@ export async function PUT(req: NextApiRequest, res: NextApiResponse<ResponseData
   }
 }
 
-export async function DELETE(req: NextApiRequest, res: NextApiResponse<ResponseData>) {
+export async function DELETE(req: NextRequest, res: NextApiResponse<ResponseData>) {
   await dbConnect();
 
-  const { id } = req.query;
-
+  const data = req.nextUrl.searchParams;
+  const id = data.get("id");
+  
   if (typeof id !== "string") {
     return res.status(400).json({ success: false, message: "Invalid ID" });
   }
