+++ src/components/BookingForm/AutoCarePackagesStep/AutocarePackages.jsx
@@ -1,17 +1,17 @@
 import { Box, Typography } from "@mui/material";
 import React, { useState, useEffect } from "react";
 
-import { AutoTabContainer } from "../../mui/AutoCarePkgs";
-import { AutoTabPackage } from "../../../app/subscribe/page";
 import { packages } from "../../../app/autocare/data";
 import { useTheme } from "../../../contexts/themeContext";
 import useMultiStepForm from "../../../hooks/useMultiStepForm";
 import { useValidation } from "../../../contexts/ValidationContext";
-import { borderRadius, display, fontFamily, lineHeight, positions } from "@mui/system";
 import Image from "next/image";
 import bg from "../../../../public/voor1.jpg";
 import CheckMark from "../../../../public/bookingFormIcons/CheckMark.svg";
 import { AutoCarePackageSubheading, BookingFormSubHeading } from "../../mui/BookingFormPackages";
+import { AutoTab, AutoTabList } from "../../mui/AutoCarePkgs";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { faCheckCircle, faChevronDown, faChevronUp } from "@fortawesome/free-solid-svg-icons";
 
 const colors = ["#087300", "#005BAC", "#BA8B1D"];
 
@@ -155,3 +155,195 @@ const AutocarePackagesCard = ({ description, price, packageType, descriptionItem
     </Box>
   );
 };
+
+
+export const AutoTabPackage = ({ pkg, index, color, onClick = undefined }) => {
+  const [additional, setAdditional] = useState(false);
+  const [duration, setDuration] = useState(false);
+  const [frequency, setFrequency] = useState(false);
+
+  return (
+      <AutoTab sx={{ width: '100%', maxWidth: '37rem' }}>
+          <div className="tab__side tab__side--front" style={{ position: "relative" }} onClick={onClick}>
+              <div className={`tab__picture tab__picture--${index + 1}`}></div>
+              <Typography className="heading">
+                  <span className={`heading--span heading--span-${index + 1}`}>{pkg.name}</span>
+              </Typography>
+              <Box sx={{ marginBottom: "8px" }}>
+                  <Typography sx={{ color: "primary.text1 !important", fontSize: "1.2rem !important", textAlign: "center", transform: "translateY(10px)" }}>
+                      FROM
+                  </Typography>
+                  <Typography sx={{ fontSize: "4rem !important", textAlign: "center", fontWeight: "900" }}>
+                      {pkg.price}
+                  </Typography>
+                  <Typography sx={{ color: "primary.text1 !important", fontSize: "1.5rem !important", textAlign: 'center' }}>
+                      {pkg.duration}
+                  </Typography>
+              </Box>
+
+              <AutoTabList sx={{
+                  padding: "8px 0",
+                  margin: "0 2rem",
+                  width: "100%",
+                  minHeight: "210px", height: "100%",
+                  justifyContent: "flex-start",
+              }}>
+                  {pkg.packages.map((item, index) => (
+                      <Box key={item} sx={{ width: "100%", display: 'flex', justifyContent: "", padding: '2px 0' }}>
+                          <FontAwesomeIcon
+                              icon={faCheckCircle}
+                              style={{
+                                  color: color,
+                                  marginRight: "1rem",
+                                  transform: "translateY(2px)",
+                              }}
+                          />
+                          <Typography sx={{ textAlign: "left !important" }}>
+                              {item}
+                          </Typography>
+                      </Box>
+                  ))}
+              </AutoTabList>
+
+              {pkg.durationOptions && onClick === undefined && (
+                  <Box sx={{ borderTop: "1px solid #00000020", margin: "0 2rem", padding: "2rem 0 1rem" }}>
+                      <Typography onClick={() => setDuration(!duration)} sx={{ fontSize: "1.5rem", textAlign: "center", fontWeight: "900" }}>
+                          Duration options
+                          {duration ? <FontAwesomeIcon
+                              icon={faChevronUp}
+                              style={{ marginLeft: "6px" }}
+                          />
+                              : <FontAwesomeIcon
+                                  icon={faChevronDown}
+                                  style={{ marginLeft: "6px" }}
+                              />
+                          }
+                      </Typography>
+                      {duration && (
+                          <AutoTabList sx={{
+                              padding: "8px 0",
+                              margin: "0",
+                              width: "100%"
+                          }}>
+                              {pkg.durationOptions?.map((item) => (
+                                  <Box key={item.duration} sx={{ width: "100%", display: 'flex', justifyContent: "", padding: "4px 0" }}>
+                                      <Typography sx={{
+                                          textAlign: "left !important",
+                                          textWrap: "nowrap",
+                                      }}>
+                                          {item.duration}
+                                      </Typography>
+                                      <Typography sx={{ textAlign: "right !important", width: "100%", fontWeight: "bold !important", color: "primary.accent" }}>
+                                          {item.additionalCost === 0 ? "" : `+ `}
+                                          € {item.additionalCost.toFixed(1)}
+                                      </Typography>
+                                  </Box>
+                              ))}
+                          </AutoTabList>
+                      )}
+                  </Box>
+              )}
+
+              {pkg.cleaningFrequencyOptions && onClick === undefined && (
+                  <Box sx={{ borderTop: "1px solid #00000020", margin: "0 2rem", padding: "2rem 0 1rem" }}>
+                      <Typography onClick={() => setFrequency(!frequency)} sx={{ fontSize: "1.5rem", textAlign: "center", fontWeight: "900" }}>
+                          Cleaning Frequency
+                          {frequency ? <FontAwesomeIcon
+                              icon={faChevronUp}
+                              style={{ marginLeft: "1rem" }}
+                          />
+                              : <FontAwesomeIcon
+                                  icon={faChevronDown}
+                                  style={{ marginLeft: "6px" }}
+                              />
+                          }
+                      </Typography>
+                      {frequency && (
+                          <AutoTabList sx={{
+                              padding: "8px 0",
+                              margin: "0",
+                              width: "100%"
+                          }}>
+                              {pkg.cleaningFrequencyOptions?.map((item) => (
+                                  <Box key={item.frequency} sx={{ width: "100%", display: 'flex', justifyContent: "space-between", padding: '4px 0', textWrap: "nowrap" }}>
+                                      <Typography sx={{
+                                          textAlign: "left !important"
+                                      }}>
+                                          {item.frequency}
+                                      </Typography>
+                                      <Typography sx={{ textAlign: "right !important", color: "primary.accent", fontWeight: "bold", textWrap: 'nowrap' }}>
+                                          {item.additionalCost === 0 ? "" : `+ `}
+                                          € {item.additionalCost.toFixed(1)}
+                                      </Typography>
+                                  </Box>
+                              ))}
+                          </AutoTabList>
+                      )}
+                  </Box>
+              )}
+              {
+                  onClick === undefined && (
+                      <>
+
+              {pkg.additionalOptions.length > 0 ? (
+                  <Box sx={{ borderTop: "1px solid #00000020", margin: "0 2rem", padding: "2rem 0 1rem" }}>
+                      <Typography onClick={() => setAdditional(!additional)} sx={{ fontSize: "1.5rem", textAlign: "center", fontWeight: "900" }}>
+                          Additional Options
+                          {additional ? <FontAwesomeIcon
+                              icon={faChevronUp}
+                              style={{ marginLeft: "1rem" }}
+                          />
+                              : <FontAwesomeIcon
+                                  icon={faChevronDown}
+                                  style={{ marginLeft: "6px" }}
+                              />
+                          }
+                      </Typography>
+
+                      {
+                          additional && (
+                              <AutoTabList sx={{
+                                  padding: "8px 0",
+                                  margin: "0",
+                                  width: "100%"
+                              }}>
+                                  {pkg.additionalOptions?.map((item) => (
+                                      <Box key={item.option} sx={{ width: "100%", display: 'flex', justifyContent: "space-between", padding: "4px 0" }}>
+                                          <Typography sx={{
+                                              textAlign: "left"
+                                          }}>
+                                              {item.option}
+                                          </Typography>
+                                          <Typography sx={{ textAlign: "right", color: "primary.accent", fontWeight: "bold", textWrap: 'nowrap' }}>
+                                              {item.additionalCost === 0 ? "" : `+ `}
+                                              € {item.additionalCost.toFixed(1)}
+                                          </Typography>
+                                      </Box>
+                                  ))}
+                              </AutoTabList>
+                          )
+                      }
+                  </Box>
+              ) : (
+                  <Box sx={{ borderTop: "1px solid #00000020", margin: "0 2rem", padding: "2rem 0 1rem" }}>
+                      <Typography sx={{ fontSize: "1.5rem", textAlign: "center", fontWeight: "900", color: '#00000050' }}>
+                          Additional Options
+                          {additional ? <FontAwesomeIcon
+                                  icon={faChevronUp}
+                                  style={{ marginLeft: "1rem" }}
+                              />
+                              : <FontAwesomeIcon
+                                  icon={faChevronDown}
+                                  style={{ marginLeft: "6px" }}
+                              />
+                          }
+                      </Typography>
+                  </Box>
+              )}
+                      </>
+                  )
+              }
+          </div>
+      </AutoTab>
+  )
+}
