+++ src/components/BookingForm/SubscriptionPackagesStep/SubscriptionPackages.jsx
@@ -1,13 +1,10 @@
 import { Box, Typography } from "@mui/material";
 import React, { useState, useEffect } from "react";
 
-import { AutoTabContainer } from "../../../components/mui/AutoCarePkgs";
-import { AutoTabPackage } from "../../../app/subscribe/page";
 import { packages } from "../../../app/subscribe/data";
 import { useTheme } from "../../../contexts/themeContext";
 import useMultiStepForm from "../../../hooks/useMultiStepForm";
 import { useValidation } from "../../../contexts/ValidationContext";
-import { borderRadius, display, fontFamily, lineHeight, positions } from "@mui/system";
 import Image from "next/image";
 import bg from "../../../../public/voor1.jpg";
