+++ src/components/ServiceDrawer/SubscriptionPackagesStep/SubscriptionPackages.jsx
@@ -2,11 +2,11 @@ import { Box } from "@mui/material";
 import React, { useState, useEffect } from "react";
 
 import { AutoTabContainer } from "../../../components/mui/AutoCarePkgs";
-import { AutoTabPackage } from "../../../app/subscribe/page";
 import { packages } from "../../../app/subscribe/data";
 import { useTheme } from "../../../contexts/themeContext";
 import useMultiStepForm from "../../../hooks/useMultiStepForm";
 import { useValidation } from '../../../contexts/ValidationContext';
+import { AutoTabPackage } from "../../BookingForm/AutoCarePackagesStep/AutocarePackages";
 
 const SubscriptionPackages = () => {
     const { theme } = useTheme();
