+++ src/components/mui/FormPkgs.jsx
@@ -109,8 +109,8 @@ export const CustomSelect = ({ name, label, options, value, onChange, ...props }
         <MenuItem value="">
           <em>None</em>
         </MenuItem>
-        {options.map((option) => (
-          <MenuItem value={option.value}>{option.label}</MenuItem>
+        {options.map((option, index) => (
+          <MenuItem key={index} value={option.value}>{option.label}</MenuItem>
         ))}
       </Select>
     </FormControl>
