+++ src/contexts/themeContext.js
@@ -1,12 +1,10 @@
 "use client";
 import React, { createContext, useContext, useState } from "react";
-import { Poppins } from "next/font/google";
 import {
   createTheme,
   ThemeProvider as MuiThemeProvider,
 } from "@mui/material/styles";
 
-const poppins = Poppins({ subsets: ["latin"], weight: ["300", "500", "700"] });
 
 const lightTheme = createTheme({
   palette: {
